refactor: 컴포넌트 내 global, page 폴더 분류 추가  및 실제 페이지를 담는 폴더는 page에서 pages로 변경
