+++ src/components/global/button/SelectButton/SelectButton.styles.ts
@@ -1,7 +1,7 @@
 // SelectButton.styles.ts
 import styled from "styled-components";
-import { colors } from "../../../style/color";
-import theme from "../../../style/theme";
+import { colors } from "../../../../style/color";
+import theme from "../../../../style/theme";
 
 export const ButtonContainer = styled.div`
   display: flex;
