+++ src/components/global/input/CustomInput/CustomInput.stories.tsx
@@ -1,7 +1,7 @@
 // CustomInput.stories.tsx
 import type { Meta, StoryObj } from "@storybook/react";
 import { CustomInput } from "./index";
-import { CustomButton } from "../../button/CustomButton/index";
+import { CustomButton } from "../../global/button/CustomButton/index";
 const meta = {
   title: "Components/CustomInput",
   component: CustomInput,
