+++ src/components/global/layout/AppBar/AppBar.styles.ts
@@ -1,5 +1,5 @@
-import styled from 'styled-components';
-import { AppBarBack } from '../../../assets/svg';
+import styled from "styled-components";
+import { AppBarBack } from "../../../../assets/svg";
 
 export const StyledHeader = styled.header`
   position: fixed;
@@ -26,6 +26,6 @@ export const Title = styled.h1<{ size: string }>`
 `;
 
 export const StyledAppBarBack = styled(AppBarBack)`
-    height: 20px;
-    cursor: pointer;
+  height: 20px;
+  cursor: pointer;
 `;
