+++ src/components/global/modal/CustomModal/CustomModal.styles.ts
