+++ src/components/global/progress/ProgressBar/ProgressBar.styles.ts
