+++ src/components/global/progress/ProgressBlock/ProgressBlock.stories.tsx
