+++ src/components/global/progress/ProgressBlock/ProgressBlock.styles.ts
