+++ src/components/global/progress/StepProgress/StepProgress.stories.tsx
