+++ src/components/global/progress/StepProgress/StepProgress.styles.ts
