+++ src/components/global/progress/StepProgress/StepProgress.tsx
