+++ src/components/global/progress/StepProgress/index.ts
