+++ src/pages/main/index.styles.ts
