+++ src/pages/sign-in/index.tsx
@@ -1,0 +1,1 @@
+//index.tsx
